refactor(session): split sqlite storage adapters
